Remove unary gRPC attestation code
Split Java gRPC API key interceptor into a separate file and BUILD
target.

We need to keep the proto files around because the web client library
still depends on them, though it does not actually work at the moment.
It will be cleaned up separately.

Ref #3442
