+++ Cargo.toml
@@ -35,7 +35,6 @@ members = [
   "oak_functions_sdk/tests/testing_module",
   "oak_functions_test_utils",
   "oak_core",
-  "oak_grpc_unary_attestation",
   "oak_idl",
   "oak_idl_build",
   "oak_idl_tests",
