+++ java/src/main/java/com/google/oak/functions/android/client/BUILD
@@ -27,11 +27,13 @@ android_library(
     manifest = "AndroidManifest.xml",
     resource_files = glob(["res/**"]),
     deps = [
-        "//java/src/main/java/com/google/oak/functions/client:client_android",
-        "//oak_functions/proto:abi_java_proto_lite",
+        "//java/src/main/java/com/google/oak/functions/client:api_key_interceptor",
+        "//java/src/main/java/com/google/oak/functions/client:client_noninteractive",
+        "//oak_remote_attestation_noninteractive/proto:oak_session_noninteractive_v1_java_grpc",
         "@com_google_guava_guava",
         "@com_google_protobuf//:protobuf_javalite",
         "@io_grpc_grpc_java//api",
+        "@io_grpc_grpc_java//netty",
     ],
 )
