+++ java/src/main/java/com/google/oak/functions/client/BUILD
@@ -14,27 +14,11 @@
 # limitations under the License.
 #
 
-load("@build_bazel_rules_android//android:rules.bzl", "android_library")
 load("@rules_java//java:defs.bzl", "java_library")
 
 package(
     default_visibility = ["//visibility:public"],
     licenses = ["notice"],
-)
-
-java_library(
-    name = "client",
-    srcs = ["AttestationClient.java"],
-    deps = [
-        "//java/src/main/java/com/google/oak/remote_attestation",
-        "//oak_functions/proto:abi_java_proto",
-        "//oak_grpc_unary_attestation/proto:server_java_grpc",
-        "//oak_grpc_unary_attestation/proto:server_java_proto",
-        "@com_google_protobuf//:protobuf_javalite",
-        "@io_grpc_grpc_java//api",
-        "@io_grpc_grpc_java//netty",
-        "@io_grpc_grpc_java//stub",
-    ],
 )
 
 java_library(
@@ -48,18 +32,10 @@ java_library(
     ],
 )
 
-# Dedicated Android target is necessary for using Tink internally.
-android_library(
-    name = "client_android",
-    srcs = ["AttestationClient.java"],
+java_library(
+    name = "api_key_interceptor",
+    srcs = ["ApiKeyInterceptor.java"],
     deps = [
-        "//java/src/main/java/com/google/oak/remote_attestation:remote_attestation_android",
-        "//oak_functions/proto:abi_java_proto_lite",
-        "//oak_grpc_unary_attestation/proto:server_java_grpc_lite",
-        "//oak_grpc_unary_attestation/proto:server_java_proto_lite",
-        "@com_google_protobuf//:protobuf_javalite",
         "@io_grpc_grpc_java//api",
-        "@io_grpc_grpc_java//okhttp",
-        "@io_grpc_grpc_java//stub",
     ],
 )
