+++ java/src/main/java/com/google/oak/functions/weather_lookup_client/BUILD
@@ -25,12 +25,13 @@ java_binary(
     srcs = ["Main.java"],
     main_class = "com.google.oak.functions.weather_lookup_client.Main",
     deps = [
-        "//java/src/main/java/com/google/oak/functions/client",
+        "//java/src/main/java/com/google/oak/functions/client:api_key_interceptor",
         "//java/src/main/java/com/google/oak/functions/client:client_noninteractive",
         "//oak_remote_attestation_noninteractive/proto:oak_session_noninteractive_v1_java_grpc",
         "@com_google_protobuf//:protobuf_java",
         "@com_google_protobuf//:protobuf_java_util",
         "@io_grpc_grpc_java//api",
+        "@io_grpc_grpc_java//netty",
         "@io_grpc_grpc_java//protobuf",
         "@io_grpc_grpc_java//stub",
     ],
