--- oak_grpc_unary_attestation/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "oak_grpc_unary_attestation"
-version = "0.1.0"
-authors = ["Juliette Pretot <julsh@google.com>"]
-edition = "2018"
-license = "Apache-2.0"
-
-[dependencies]
-oak_remote_attestation = { path = "../oak_remote_attestation" }
-oak_remote_attestation_sessions = { path = "../oak_remote_attestation_sessions" }
-oak_remote_attestation_sessions_client = { path = "../oak_remote_attestation_sessions_client" }
-anyhow = "*"
-prost = "*"
-serde = { version = "*", features = ["derive"] }
-tokio = { version = "*", features = [
-  "fs",
-  "macros",
-  "process",
-  "sync",
-  "rt-multi-thread"
-] }
-tonic = "*"
-tower = "*"
-async-trait = "*"
-
-[build-dependencies]
-oak_grpc_utils = { path = "../oak_grpc_utils" }
